bugfixes/enhancements
- general: added static type checkers for templated callbacks
- memory: added variative callback signatures for 'perform*'
- memory: added push/pop source line tracking
- memory: improved debug error controls
- memory: AllocRef and PoolRef now will not cause use-after-free
- event: fixed ALooper engine
- thirdparty: libbacktrace now packaged as internal dependency

-

+++ core/memory/pool/SPMemPoolPool.cc
@@ -483,8 +483,14 @@ void initialize() {
 
 void terminate() {
 	if (s_global_init.fetch_sub(1) == 1) {
-		Pool::destroy(s_global_pool);
-		delete s_global_allocator;
+		if (s_global_pool) {
+			Pool::destroy(s_global_pool);
+			s_global_pool = nullptr;
+		}
+		if (s_global_allocator) {
+			delete s_global_allocator;
+			s_global_allocator = nullptr;
+		}
 	}
 }
